+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Table of contents to Project Settings Page (only shows h2 headings)
 - OG Image generation for generated posts
+- Handle link of an image in submit blog post endpoint
 
 ### Fixed
 - Cloudlfare Turnstyle now actaully does stuff